requirements.txt
# MLB Analytics System - Complete Dependency List
# Core Python requirements for honest MLB prediction system

# Database and ORM
sqlalchemy==2.0.46
psycopg2-binary==2.9.11

# Data Science and Machine Learning
pandas==3.0.0
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.0
xgboost==3.1.3

# Environment and Configuration
python-dotenv==1.2.1

# Date/Time handling
python-dateutil==2.9.0.post0

# Data Collection (MLB APIs)
requests>=2.28.0
mlb-statsapi>=1.6.0

# Additional ML/Data dependencies
joblib==1.5.3
threadpoolctl==3.6.0

# Database connection pooling
greenlet==3.3.1

# Type hints (for older Python versions)
typing-extensions==4.15.0

# Utility
six==1.17.0

# Development/Optional (uncomment if needed)
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0
# plotly>=5.0.0
# beautifulsoup4>=4.11.0
# lxml>=4.9.0